lab2src/requirements.txt
torch 
torchaudio
torchvision
tensorflow 
setuptools
wheel
tokenizers
transformers
streamlit
python-dotenv
pyyaml